+++ kit/dapp/messages/en.json
@@ -2566,13 +2566,11 @@
       "approve": "Approve",
       "approve-xvp-settlement": "Approve XvP Settlement",
       "approve-xvp-settlement-description": "Grant allowances to the XvP Settlement contract for the assets that will be transferred from your wallet.",
-      "approve-description": "The XvP Settlement contract will be granted the following approvals. Then the settlement will be marked as approved by you.",
+      "approve-description": "The XvP Settlement contract will be granted approval to transfer the assets configured in the settlement on your behalf. Then the settlement will be marked as approved by you.",
       "asset": "Asset",
-      "asset-description": "The asset that will be transferred from your wallet.",
       "amount": "Amount",
-      "amount-description": "The amount of the asset that will be transferred.",
       "to": "To",
-      "to-description": "The wallet address of the recipient of the asset."
+      "xvp-settlement": "XvP Settlement"
     }
   },
   "transactions": {
